PO-1696: Add GET Defendant Account Enforcement Status endpoint.

+++ src/main/java/uk/gov/hmcts/opal/dto/legacy/common/EnforcerReference.java
@@ -23,7 +23,7 @@ public class EnforcerReference implements ToXmlString {
 
     @JsonProperty(value = "enforcer_id", required = true)
     @XmlElement(name = "enforcer_id", required = true)
-    private Integer enforcerId;
+    private Long enforcerId;
 
     @JsonProperty(value = "enforcer_name", required = true)
     @XmlElement(name = "enforcer_name", required = true)
